+++ packages/sources/tiingo/src/endpoint/common/iex-router.ts
@@ -8,8 +8,8 @@ import { httpTransport } from '../http/iex'
 import { wsTransport } from '../ws/iex'
 
 const inputParameters = new InputParameters({
-  ticker: {
-    aliases: ['base', 'from', 'coin'],
+  base: {
+    aliases: ['ticker', 'from', 'coin'],
     required: true,
     type: 'string',
     description: 'The stock ticker to query',
@@ -28,7 +28,7 @@ export const endpoint = new AdapterEndpoint({
   transportRoutes: new TransportRoutes<IEXEndpointTypes>()
     .register('ws', wsTransport)
     .register('rest', httpTransport),
-  defaultTransport: 'ws',
+  defaultTransport: 'rest',
   inputParameters: inputParameters,
   overrides: overrides.tiingo,
 })